refactor: moved compiler errors out of compiler.h
